refactor: exported front end data validator in their own file
